+++ elide-standalone/src/main/java/com/yahoo/elide/standalone/config/ElideResourceConfig.java
@@ -5,6 +5,8 @@
  */
 package com.yahoo.elide.standalone.config;
 
+import com.codahale.metrics.MetricRegistry;
+import com.codahale.metrics.health.HealthCheckRegistry;
 import com.yahoo.elide.Elide;
 import com.yahoo.elide.ElideSettings;
 import com.yahoo.elide.core.DataStore;
@@ -35,6 +37,9 @@ public class ElideResourceConfig extends ResourceConfig {
     private final ServiceLocator injector;
 
     public static final String ELIDE_STANDALONE_SETTINGS_ATTR = "elideStandaloneSettings";
+
+    private static MetricRegistry metricRegistry = null;
+    private static HealthCheckRegistry healthCheckRegistry = null;
 
     /**
      * Constructor
@@ -96,5 +101,21 @@ public class ElideResourceConfig extends ResourceConfig {
      */
     private void registerFilters(List<Class<?>> filters) {
         filters.forEach(this::register);
+    }
+
+    public static MetricRegistry getMetricRegistry() {
+        if (metricRegistry == null) {
+            metricRegistry = new MetricRegistry();
+        }
+
+        return metricRegistry;
+    }
+
+    public static HealthCheckRegistry getHealthCheckRegistry() {
+        if (healthCheckRegistry == null) {
+            healthCheckRegistry = new HealthCheckRegistry();
+        }
+
+        return healthCheckRegistry;
     }
 }